Started to design and implement the Driver mechanism, starting with the Syscalls.  Syscalls require an ISR, as do a lot of Drivers, so put in a configuration structure / parsing for that.  Came to the realisation that the whole ISR / vector table implementation could be significantly improved by leveraging the same Config-based discovery, so rejigged that into something much cleaner and more easily extensible.  Plus the dependency on weak symbols and name-manging has been dropped.
Along the same lines, Config-based discovery can be (and will be) used for creating the Task Stacks, but these thoughts have merely been recorded as TODOs.

Normalised some naming, specifically around task configuration; IHaveFoo is for concepts evaluating a Foo member, whereas IFoo is a concept evaluating against a Foo.

Realised that names beginning with an underscore followed by an uppercase letter are reserved _in any scope_ by the C++ standard, so renamed the _SomethingPrivate utility classes to _$SomethingPrivate.  Once modules are mature enough in g++ to be introduced into the codebase then this convention of 'hiding' details in the headers can largely disappear.

Added some simple documentation in the form of a few basic README.md files.  These will need elaborating somewhat, but they're merely intended as aide-memoires covering salient details.

Everything so far is still very much a work-in-progress as concepts and implementation details are worked through.