Another few hundred lines of code just to populate 4 bytes in a table - abstraction FTW ! :-/
Introduced trampolines for the ISRs, meaning that any per-core ISRs (ie. non-naked) don't need to be decorated with platform-specific attributes because they're attached to the trampoline instead.

Added McuCoreTraits to allow executing code to query which MCU core it is running on, which is initially leveraged to look up and dispatch the per-core ISRs.

Added the NakedToPerCoreIsrAdapter which provides a naked ISR with per-core storage for ISR instances and does the lookup / dispatch.

The Syscall driver has been converted from a naked ISR to a per-core ISR to simplify the onward dispatch.

This is clearly a WIP commit but it is at the point where everything compiles to what looks like correct assembly, even if there are several hacks and warts.  Added various TODOs to indicate next steps / direction.