There are difficulties with the *Apis containers around construction, testing and test doubles.  Conceptually the explicit *Apis containers are only required to facilitate (easier) dependency injection.
C++ does not (yet) have reflection, and without the ability to take the address of a constructor to determine its arguments, it seems dependency injection without knowing types up-front is not possible.  However, after a few days of hacking and researching, it appears as though there is a 'work-around'.

This commit is the first part of a proper DI implementation for the kernel, tailored for static allocation and resource-constrained architectures.  The ConstructorParameters class encapsulates the badness of stateful metaprogramming in the absense of proper reflection.  Once DI has been introduced then the *Apis containers can be removed and the need for the various *Factory classes can be revisited with an eye to removal and simplification.